Move Payment Transaction flat file.
